+++ js/util.js
@@ -1,0 +1,17 @@
+function getRandomNumber(from, to) {
+  //TODO: упростить return для критерия Д18
+  if (to <= from && from >= 0 && to >= 0) {
+    return new Error('переданный параметр не соответсвует условию');
+  }
+  const random = from + Math.random() * (to + 1 - from);
+  return Math.floor(random);
+}
+getRandomNumber(0, 20);
+
+
+function isSuitableStringLength(line, maxLine) {
+  return line.length <= maxLine;
+}
+isSuitableStringLength('test', 5);
+
+export { getRandomNumber };
